// client/src/app/models/course.model.ts
export class Course {
  id: string;
  title: string;
  grade: string;
  desc: string;
  type: string;
  project_id: string;
  project_name: string;
  teacher_id: string;
  teacher_name: string;
  student_id: string;
  student_name: string;
  start: any;
  end: any;
  start_unix: any;
  end_unix: any;
  booked: boolean;
}